Docker/src/requirements.txt
selenium==3.141.0
mysql-connector-python==8.0.23
